plot tile
