+++ explore/src/main/scala/explore/plots/ObjectPlotData.scala
@@ -1,7 +1,7 @@
 // Copyright (c) 2016-2023 Association of Universities for Research in Astronomy, Inc. (AURA)
 // For license information see LICENSE or https://opensource.org/licenses/BSD-3-Clause
 
-package explore.targeteditor.plots
+package explore.plots
 
 import cats.Eq
 import cats.Order
